backend/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
Pillow==10.0.0
numpy==1.25.0
tensorflow==2.11.0
gunicorn==20.1.0
